Add files via upload
The objective of the project is to predict the selling prices of the houses based on some factors such as location, total area, etc. I used the Bangalore House Prices dataset for my prediction. I have done the data load and cleaning, feature engineering, dimensionality reduction, outlier detection and removal, one-hot encoding, etc. I built a Linear Regression model with an accuracy of 87%